move debug code into subpackage of debugging
