+++ test/debugging/jsonclientcaller/HandleManagerTest.java
@@ -1,4 +1,4 @@
-package debugging;
+package debugging.jsonclientcaller;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
