+++ test/debugging/jsonclientcaller/HandleServiceTest.java
@@ -1,4 +1,4 @@
-package debugging;
+package debugging.jsonclientcaller;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
